requirements.txt
Authlib==1.1.0
boto3==1.26.13
botocore==1.29.13
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
Flask==2.2.2
google-auth==2.14.1
google-auth-oauthlib==0.7.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
pathlib==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
vendor==0.1
Werkzeug==2.2.2
zipp==3.10.0